+++ scripts/filter.rb
@@ -6,6 +6,13 @@ names = {}
 
 puts "Filtering names.csv to names_filtered.csv\n\n"
 
+# filter_ok returns true if conditions to keep a name are satisfied.
+def filter_ok(row)
+  curated = %w[Curated AutoCurated].include?(row['Curation'])
+  high_odds = row['OccurrencesNumber'].to_i > 1 && row['OddsLog10'].to_f > 6.0
+  curated || high_odds
+end
+
 # copy headers of names first
 r = File.open('names.csv')
 w = File.open('names_filtered.csv', 'w:utf-8')
@@ -22,7 +29,7 @@ saved = 0
 r.each do |l|
   count += 1
   puts "Traversed #{count} names, saved #{saved}" if (count % 1_000_000).zero?
-  next unless %w[Curated AutoCurated].include?(l['Curation']) || l['OddsLog10'].to_f > 6.0
+  next unless filter_ok(l)
 
   w << l
   names[l['NameID']] = true